Make clone_dataset() patchable and less monolithic
This change is equivalent to the commuluative effort posted at
https://github.com/datalad/datalad-next/pull/99. Its description
is included below.

In addition to this original changeset:

- RIA-related helper are moved into `clone_ria.py` too
- internal patches `ria` and `ephemeral` are applied via functions
  that can be easily patched-out, in order to turn the patches off
  externally -- for example when an extension starts to provide a
  superior replacement.

The (adjusted for filenames) description of the original PR follows:

The series provides a patch to replace `clone_dataset()` the key
workhorse of datalad-core's `clone` command. The patch causes no change
to any functionality, but refactors the code from large monolithic
functions into significantly smaller, more comprehensible units.
Emphasis is made on separating function that control the flow of
processing (`clone.py`) from utilities that perform focused
tasks (`clone_utils.py`).

Key feature of the refactoring is the provisioning of a set of dedicated
function that implement particular stages of the clone and post-clone
setup processing:

- `_try_clone_candidates()` (the main clone attempt loop)
- `_try_clone_candidate()` (handle a single clone attempt)
- `_try_git_clone_candidate()` (only workhorse ATM with all git-clone specifics)
- `_post_gitclone_processing_()` (implements the logic to call all others)
- `_post_git_init_processing_()`
- `_pre_annex_init_processing_()`
- `_annex_init()`
- `_post_annex_init_processing_()`
- `_pre_final_processing_()`

All functions implement a keyword-argument-only API. All functions
ending with the name `processing_` are implemented as generators that
yield standard DataLad result records.

Any of these functions can be "patched" following the pattern used in
`datalad-next` for any kind of modifications of datalad-core. This
capability is demonstrated by pulling out two optional features,
originally intermingled with essential code:

- `clone_ria.py`
- `clone_ephemeral.py`
  (related: https://github.com/datalad/datalad/issues/6961)

(Note that for both patches, the documentation is not patched right now,
because it seems unnecessary, and can be left to when the patch is
turned off externally)

These document how it is possible, with this new implementation, to
provide such add-on features externally, without having to introduce
them in datalad-core directly. They also show, how the patches can
benefit from the keyword-argument-only API. They should continue to work
as designed, as long as the particular keyword-arguments the require
continue to be provided with the same semantics.

Taken together, this series provides a solution to a number of issues
that are linked below with a brief comment:

- https://github.com/datalad/datalad/issues/4903 -- more-or-less
  arbitrary modifications can be implemented following `git-clone`. The
  latter is itself encapsulated in the `_try_clone_candidate()` utility,
  so it would even be possible to patch-in means to establish repository
  clones that do not involve calling `git-clone`

- This is related to
  https://github.com/datalad/datalad-registry/issues/7 -- while we
  already have flexibility to add handling of new URL-types (e.g., URL
  mapping, subdataset candidate URL configurations), now, either
  `_try_clone_candidates()` or another utility
  (`_generate_candidate_clone_sources()`) can be patched to query an
  external resolver of some kind, including a DataLad registry

- https://github.com/datalad/datalad/issues/5289 -- the datalad-next
  style patching has proven to be somewhat reliable, and continue to
  grow in usage. This series modifies datalad-core to enable more surgical
  patching for add-on functionality. Continuing this for other parts of
  the code base would lessen the need for (possibly expensive) hooks that
  may require additional setup and maintenance.

The patch-in-a-clone-feature approach should be applicable to the
growling list of open issues asking for additional (special interest)
capabilities. Examples:

- https://github.com/datalad/datalad/issues/6408
- https://github.com/datalad/datalad/issues/5094
- https://github.com/datalad/datalad/issues/4142

Lastly, this approach can help to expedite work on rejuvenating datalad
components that struggle to keep up within the corset of datalad-core,
with its ambition to remain as stable as possible. One example is the
RIA functionality that alone is associated with 7% of the open issues,
and would benefit from a few faster, less constrained development
cycles. The previous impossibility of `clone` to be customized
externally (like done here), was one of the main reasons for having to
introduce RIA in datalad-core directly, rather than having it prove
itself in a dedicated extension.
